- Define bower directory and refactor config

+++ config/init.js
@@ -6,25 +6,45 @@ import User from '../app/models/User';
 
 dotenv.config();
 
-const db = process.env.MONGO_TEST;
+const environment = process.env.NODE_ENV;
+let db;
+
+switch(environment) {
+  case 'test':
+    db = process.env.MONGO_TEST;
+    break;
+  case 'dev':
+    db = process.env.MONGO_URL;
+    break;
+}
 
 mongoose.connect(db);
 
 mongoose.connection.once('connected', () => {
-  mongoose.connection.db.dropCollection('users', (err) => {
-    console.log('Dropped User');
-    const salt = bcrypt.genSaltSync(10);
 
-    User.create({
-      name: 'Super Admin',
-      email: process.env.SUPEREMAIL,
-      password: bcrypt.hashSync(process.env.SUPERPASSWORD, salt),
-      role: 'super',
-      phone: process.env.SUPERPHONE,
-      department: 'Management'
-    }, (err, user) => {
-      console.log('Created Super User');
-      process.exit(0);
+  if (environment == 'test') {
+    mongoose.connection.db.dropCollection('users', (err) => {
+      console.log('Dropped User');
     });
+  }
+
+  const salt = bcrypt.genSaltSync(10);
+
+  User.findOne({ email: process.env.SUPEREMAIL}, (err, user) => {
+    if (!user) {
+      User.create({
+        name: 'Super Admin',
+        email: process.env.SUPEREMAIL,
+        password: bcrypt.hashSync(process.env.SUPERPASSWORD, salt),
+        role: 'super',
+        phone: process.env.SUPERPHONE,
+        department: 'Management'
+      }, (err, user) => {
+        console.log('Created Super User');
+        process.exit(0);
+      });
+    } else {
+      process.exit(0);
+    }
   });
 });
